Move proptest to dev-dependencies
By getting rid of /tests and just putting everything under #[cfg(test)]
