+++ sigma_fun/Cargo.toml
@@ -13,13 +13,13 @@ secp256kfun = { path = "../secp256kfun", version = "^0.3.2-alpha.0", default-fea
 curve25519-dalek = { version = "3", default-features = false, optional = true, features = ["u64_backend"] }
 rand_chacha = "0.2"
 serde_crate = { package = "serde", version = "1.0", optional = true, default-features = false, features = ["derive"] }
-proptest = "0.10"
 
 [dev-dependencies]
 secp256kfun = { path = "../secp256kfun", version = "^0.3.2-alpha.0", default-features = false, features = ["proptest"] }
 rand = "0.7"
 sha2 = "0.9"
 bincode = "1"
+proptest = "0.10"
 
 [features]
 default = ["secp256k1", "ed25519", "alloc"]
